Fix errors and remove extra files and folders from client dir
