+++ src/client/package.json
@@ -31,6 +31,7 @@
     "storybook": "^7.0.13",
     "tailwindcss": "^3.3.1",
     "typescript": "^5.0.2",
-    "vite": "^4.3.0"
+    "vite": "^4.3.0",
+    "zustand": "^4.3.8"
   }
 }
