+++ src/client/src/app.tsx
@@ -1,15 +1,11 @@
-import CleanApp from "@pages/clean-app-page";
-import Testing from "@pages/testing";
+import Feed from "@pages/feed";
 import { Route } from "wouter-preact";
 
 export function App() {
   return (
     <>
       <Route path="/">
-        <CleanApp />
-      </Route>
-      <Route path="/testing">
-        <Testing />
+        <Feed />
       </Route>
     </>
   );
