--- src/client/src/features/testing/Testing.tsx
@@ -1,7 +1,0 @@
-export default function Testing() {
-  return (
-    <>
-      <div> This is testing</div>
-    </>
-  );
-}
