--- src/client/src/pages/clean-app-page/App.tsx
@@ -1,12 +1,0 @@
-import Testing from "@features/testing";
-export default function App() {
-  return (
-    <>
-      <div className="bg-red-500 h-screen w-screen m-auto flex">
-        <div className="w-fit h-fit text-white text-2xl m-auto pb-32">
-          Clean App
-        </div>
-      </div>
-    </>
-  );
-}
