--- src/client/src/pages/clean-app-page/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./App.tsx";
