+++ src/client/src/pages/feed/Feed.tsx
@@ -1,0 +1,9 @@
+export default function Feed() {
+  return (
+    <>
+      <div className="bg-white-500 h-screen w-screen m-auto flex">
+        Under construction
+      </div>
+    </>
+  );
+}
