+++ src/client/src/pages/feed/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Feed.js";
