--- src/client/src/pages/profile/Profile.tsx
@@ -1,3 +1,0 @@
-export default function ProfilePage() {
-  return <div>Profile Page</div>;
-}
