--- src/client/src/pages/testing/Testing.tsx
@@ -1,9 +1,0 @@
-export default function Testing() {
-  return (
-    <>
-      <div className="bg-site-backdrop h-screen w-screen m-auto flex">
-        <div className="w-fit h-fit text-slate-900 text-2xl m-auto border-slate-900 border-2 py-32 px-32"></div>
-      </div>
-    </>
-  );
-}
