--- src/client/src/pages/testing/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Testing.tsx";
