+++ src/client/src/routes/nested-router/index.ts
